pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "elftriage"
version = "0.1.0"
description = "Static analysis triage tool for x86_64 ELF binaries"
requires-python = ">=3.11"
dependencies = [
    "pyelftools>=0.31",
    "capstone>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=24.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

[project.scripts]
elftriage = "elftriage.cli:main"

[tool.black]
line-length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "capstone"
ignore_missing_imports = true
